🍎 Chore: 멀티모듈 추가 구성 진행
