+++ semo-batch-retention/build.gradle
@@ -1,0 +1,14 @@
+dependencies {
+	implementation project(':semo-core')
+
+	implementation 'org.springframework.boot:spring-boot-starter-batch'
+	testImplementation 'org.springframework.batch:spring-batch-test'
+}
+
+tasks.bootJar {
+	enabled = false
+}
+
+tasks.jar {
+	enabled = true
+}
