+++ semo-batch-retention/src/main/resources/application.yml
@@ -13,9 +13,11 @@ spring:
       enabled: false
 
 server:
-  port: 9090
+  port: 9091
 
 logging:
   level:
     org.hibernate.type.descriptor.sql: trace
 
+backup:
+  path: ./backup/session-data
