+++ semo-batch-retention/src/main/resources/banner.txt
@@ -1,0 +1,7 @@
+       *    *       *    *       *    *      *           *    *          *        *
+ ,-*-. ,-*----.,-*.   ,**. ,--**-.  *     ,-----. * ,--*. ,**-****-.,---**.,*-.  ,--.
+'   .-'|* .-*-'|   `*'   |' *.-.  ',--**-.|  |) /_ /  O  \'--.  .--'  .--./|  '-*'  |
+`*  `-.|  `--, | *|'.'|  ||  | | *|'-----'|  .-.  \  .*.  |  |  |  |  | *  |  .--.  |
+.-'    |  `*--.|  |   |* |'  '*'  '     * |  '*-' /  | |  |  |  |  '  '--'\|  |  |  |
+`--*--'`-*--*-'`--'   `--' `-*--*'  *     `**---*'`--' `--'  `*-'   `**---'`--'  `**'
+----------------------------------------------------------Powered by Spring Boot ${spring-boot.version}
